refactor: notification system
