+++ packages/plugin-manual-notification/src/client/NotificationConfigPane.tsx
@@ -2,8 +2,9 @@ import { ExtendCollectionsProvider, SchemaComponent } from '@tachybase/client';
 
 import { App, Card } from 'antd';
 
-import { useSendActionProps } from './components/SendNotifications';
-import { notificationCollection, notificationSchema } from './schemas/notificationConfigSchema';
+import { notificationCollection } from './collections/notification.collection';
+import { useSendActionProps } from './hooks/useSendActionProps';
+import { notificationSchema } from './schemas/schemaNotificationConfig';
 
 export const NotificationConfigPane = () => {
   return (
