+++ packages/plugin-manual-notification/src/client/schemas/createNotification.ts
@@ -1,0 +1,90 @@
+import { ISchema } from '@tachybase/schema';
+
+import { notificationCollection } from '../collections/notification.collection';
+import { NAMESPACE } from '../locale';
+import { NotificationFieldset } from './fieldsetNotification';
+
+export const createNotification: ISchema = {
+  type: 'void',
+  'x-action': 'create',
+  'x-acl-action': 'create',
+  title: "{{t('Add new')}}",
+  'x-component': 'Action',
+  'x-decorator': 'ACLActionProvider',
+  'x-component-props': {
+    openMode: 'drawer',
+    type: 'primary',
+    component: 'CreateRecordAction',
+    icon: 'PlusOutlined',
+  },
+  'x-align': 'right',
+  'x-acl-action-props': {
+    skipScopeCheck: true,
+  },
+  properties: {
+    drawer: {
+      type: 'void',
+      title: `{{ t("Add notification", { ns: "${NAMESPACE}" }) }}`,
+      'x-component': 'Action.Container',
+      'x-component-props': {
+        className: 'tb-action-popup',
+      },
+      properties: {
+        body: {
+          type: 'void',
+          'x-acl-action-props': {
+            skipScopeCheck: true,
+          },
+          'x-acl-action': `notificationConfigs:create`,
+          'x-decorator': 'FormBlockProvider',
+          'x-use-decorator-props': 'useCreateFormBlockDecoratorProps',
+          'x-decorator-props': {
+            dataSource: 'main',
+            collection: notificationCollection,
+          },
+          'x-component': 'CardItem',
+          properties: {
+            form: {
+              type: 'void',
+              'x-component': 'FormV2',
+              'x-use-component-props': 'useCreateFormBlockProps',
+              properties: {
+                actionBar: {
+                  type: 'void',
+                  'x-component': 'ActionBar',
+                  'x-component-props': {
+                    style: {
+                      marginBottom: 24,
+                    },
+                  },
+                  properties: {
+                    cancel: {
+                      title: '{{ t("Cancel") }}',
+                      'x-component': 'Action',
+                      'x-use-component-props': 'useCancelActionProps',
+                    },
+                    submit: {
+                      title: '{{ t("Submit") }}',
+                      'x-component': 'Action',
+                      'x-use-component-props': 'useCreateActionProps',
+                      'x-component-props': {
+                        type: 'primary',
+                        htmlType: 'submit',
+                      },
+                      'x-action-settings': {
+                        assignedValues: {},
+                        triggerWorkflows: [],
+                        pageMode: false,
+                      },
+                    },
+                  },
+                },
+                ...NotificationFieldset,
+              },
+            },
+          },
+        },
+      },
+    },
+  },
+};
