+++ packages/plugin-manual-notification/src/client/schemas/schemaNotificationConfig.ts
@@ -1,0 +1,184 @@
+import { ISchema } from '@tachybase/schema';
+
+import { notificationCollection } from '../collections/notification.collection';
+import { NAMESPACE } from '../locale';
+import { createNotification } from './createNotification';
+import { updateNotificationConfig } from './updateNotificationConfig';
+
+export const notificationSchema: ISchema = {
+  type: 'void',
+  properties: {
+    notificationconfig: {
+      type: 'void',
+      'x-decorator': 'TableBlockProvider',
+      'x-decorator-props': {
+        collection: notificationCollection,
+        action: 'list',
+        params: {
+          pageSize: 50,
+        },
+      },
+      title: `{{t("Notification Config", { ns: "${NAMESPACE}" })}}`,
+      properties: {
+        actions: {
+          type: 'void',
+          'x-component': 'ActionBar',
+          'x-component-props': {
+            style: {
+              marginBottom: 16,
+            },
+          },
+          properties: {
+            refresh: {
+              type: 'void',
+              title: '{{ t("Refresh") }}',
+              'x-component': 'Action',
+              'x-use-component-props': 'useRefreshActionProps',
+              'x-component-props': {
+                icon: 'ReloadOutlined',
+              },
+            },
+            delete: {
+              type: 'void',
+              title: '{{ t("Delete") }}',
+              'x-component': 'Action',
+              'x-action': 'destroy',
+              'x-decorator': 'ACLActionProvider',
+              'x-use-component-props': 'useBulkDestroyActionProps',
+              'x-component-props': {
+                icon: 'DeleteOutlined',
+                confirm: {
+                  title: "{{t('Delete')}}",
+                  content: "{{t('Are you sure you want to delete it?')}}",
+                },
+                actionCallback: '{{ dataSourceDeleteCallback }}',
+              },
+            },
+            create: createNotification,
+          },
+        },
+        table: {
+          type: 'array',
+          'x-component': 'TableV2',
+          'x-use-component-props': 'useTableBlockProps',
+          'x-component-props': {
+            rowKey: 'key',
+            rowSelection: {
+              type: 'checkbox',
+            },
+          },
+          properties: {
+            id: {
+              type: 'void',
+              'x-decorator': 'TableV2.Column.Decorator',
+              'x-component': 'TableV2.Column',
+              'x-component-props': {
+                width: 100,
+              },
+              properties: {
+                id: {
+                  type: 'string',
+                  'x-component': 'CollectionField',
+                  'x-read-pretty': true,
+                },
+              },
+            },
+            title: {
+              type: 'void',
+              'x-decorator': 'TableV2.Column.Decorator',
+              'x-component': 'TableV2.Column',
+              properties: {
+                title: {
+                  type: 'string',
+                  'x-component': 'CollectionField',
+                  'x-component-props': {
+                    ellipsis: true,
+                  },
+                  'x-read-pretty': true,
+                },
+              },
+            },
+            detail: {
+              type: 'void',
+              'x-decorator': 'TableV2.Column.Decorator',
+              'x-component': 'TableV2.Column',
+              properties: {
+                detail: {
+                  type: 'string',
+                  'x-component': 'CollectionField',
+                  'x-component-props': {
+                    ellipsis: true,
+                  },
+                  'x-read-pretty': true,
+                },
+              },
+            },
+            duration: {
+              type: 'void',
+              'x-decorator': 'TableV2.Column.Decorator',
+              'x-component': 'TableV2.Column',
+              properties: {
+                duration: {
+                  type: 'string',
+                  'x-component': 'CollectionField',
+                  'x-disabled': true,
+                  // 'x-read-pretty': true,
+                },
+              },
+            },
+            level: {
+              type: 'void',
+              'x-decorator': 'TableV2.Column.Decorator',
+              'x-component': 'TableV2.Column',
+              properties: {
+                level: {
+                  type: 'string',
+                  'x-component': 'CollectionField',
+                  'x-read-pretty': true,
+                },
+              },
+            },
+            actions: {
+              type: 'void',
+              title: '{{t("Actions")}}',
+              'x-component': 'TableV2.Column',
+              properties: {
+                actions: {
+                  type: 'void',
+                  'x-component': 'Space',
+                  'x-component-props': {
+                    // split: '|',
+                  },
+                  properties: {
+                    send: {
+                      type: 'void',
+                      title: `{{t("Send", { ns: "${NAMESPACE}" })}}`,
+                      'x-component': 'Action.Link',
+                      'x-use-component-props': 'useSendActionProps',
+                    },
+                    update: updateNotificationConfig,
+                    destroy: {
+                      type: 'void',
+                      title: '{{ t("Delete") }}',
+                      'x-action': 'destroy',
+                      'x-decorator': 'ACLActionProvider',
+                      'x-component': 'Action.Link',
+                      'x-use-component-props': 'useDestroyActionProps',
+                      'x-component-props': {
+                        icon: 'DeleteOutlined',
+                        confirm: {
+                          title: "{{t('Delete record')}}",
+                          content: "{{t('Are you sure you want to delete it?')}}",
+                        },
+                      },
+                    },
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+    },
+  },
+};
