+++ packages/plugin-manual-notification/src/client/schemas/updateNotificationConfig.ts
@@ -1,0 +1,81 @@
+import { ISchema } from '@tachybase/schema';
+
+import { notificationCollection } from '../collections/notification.collection';
+import { NAMESPACE } from '../locale';
+import { NotificationFieldset } from './fieldsetNotification';
+
+export const updateNotificationConfig: ISchema = {
+  type: 'void',
+  title: '{{ t("Edit") }}',
+  'x-action': 'update',
+  'x-component': 'Action.Link',
+  'x-component-props': {
+    openMode: 'drawer',
+    icon: 'EditOutlined',
+  },
+  'x-decorator': 'ACLActionProvider',
+  properties: {
+    drawer: {
+      type: 'void',
+      title: `{{ t("Edit notification", { ns: "${NAMESPACE}" }) }}`,
+      'x-component': 'Action.Container',
+      'x-component-props': {
+        className: 'tb-action-popup',
+      },
+      properties: {
+        card: {
+          type: 'void',
+          'x-acl-action-props': {
+            skipScopeCheck: false,
+          },
+          'x-acl-action': `${notificationCollection.name}:update`,
+          'x-decorator': 'FormBlockProvider',
+          'x-use-decorator-props': 'useEditFormBlockDecoratorProps',
+          'x-decorator-props': {
+            action: 'get',
+            dataSource: 'main',
+            collection: notificationCollection,
+          },
+          'x-component': 'CardItem',
+          properties: {
+            form: {
+              type: 'void',
+              'x-component': 'FormV2',
+              'x-use-component-props': 'useEditFormBlockProps',
+              properties: {
+                actionBar: {
+                  type: 'void',
+                  'x-component': 'ActionBar',
+                  'x-component-props': {
+                    style: {
+                      marginBottom: 24,
+                    },
+                  },
+                  properties: {
+                    cancel: {
+                      title: '{{ t("Cancel") }}',
+                      'x-component': 'Action',
+                      'x-use-component-props': 'useCancelActionProps',
+                    },
+                    submit: {
+                      title: '{{ t("Submit") }}',
+                      'x-component': 'Action',
+                      'x-use-component-props': 'useUpdateActionProps',
+                      'x-component-props': {
+                        type: 'primary',
+                      },
+                      'x-action-settings': {
+                        isDeltaChanged: true,
+                      },
+                    },
+                  },
+                },
+                ...NotificationFieldset,
+              },
+            },
+          },
+        },
+      },
+    },
+  },
+};
